Move supported servers to utils
